GPS: split the driver into per-protocol translation units
SerialGPS.cpp had grown to hold everything GPS: two wire protocols, the baud
rate detection, the u-blox configuration ladder and the CRSF telemetry. Split
the two protocols out so each unit is about one thing.

SerialGPS-nmea.cpp has the sentence framing, the checksum and the field
parsers. SerialGPS-ublox.cpp has the message framing, the NAV-PVT decode and
every configuration message sent back to the module. SerialGPS.cpp keeps the
baud detection, the detect/configure state machine and the telemetry frames.

processBytes() is now just the demultiplexer: the first byte after a message
decides which parser owns the stream, and each protocol frames its own bytes
from there. The two parsers still share the one message buffer, since only one
message is ever in flight, so this costs no RAM.

The state machine no longer builds UBX payloads inline. beginConfigure() calls
queueCfgValset() and the VALSET timeout path calls queueCfgLegacy(), both in
the u-blox unit, so no UBX message layout is left in the main file and no
detection constant leaks into a protocol unit.

No behaviour change. All 38 test_gps tests (132 in the suite) pass unchanged.
